Revert equality comparision related algorithms.
