+++ src/Exceptionless/Extensions/CollectionEqualityExtensions.cs
@@ -1,33 +1,9 @@
 ﻿using System;
-using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
 
 namespace Exceptionless {
     internal static class CollectionEqualityExtensions {
-        private static bool ElementEquals(object source, object other) {
-            if (ReferenceEquals(null, source) && ReferenceEquals(null, other)) {
-                return true;
-            }
-
-            if (ReferenceEquals(null, source) || ReferenceEquals(null, other)) {
-                return false;
-            }
-
-            if (ReferenceEquals(source, other)) {
-                return true;
-            }
-
-            if (source is IDictionary sourceDictionary && other is IDictionary otherDictionary) {
-                return sourceDictionary.OfType<DictionaryEntry>().ToDictionary(entry => entry.Key, entry => entry.Value).CollectionEquals(otherDictionary.OfType<DictionaryEntry>().ToDictionary(entry => entry.Key, entry => entry.Value));
-            }
-
-            if (source is IEnumerable sourceEnumerable && other is IEnumerable otherEnumerable) {
-                return sourceEnumerable.OfType<object>().CollectionEquals(otherEnumerable.OfType<object>());
-            }
-
-            return source.Equals(other);
-        }
         public static bool CollectionEquals<T>(this IEnumerable<T> source, IEnumerable<T> other) {
             if (ReferenceEquals(null, source) && ReferenceEquals(null, other)) {
                 return true;
@@ -50,7 +26,7 @@ namespace Exceptionless {
                     return false;
                 }
 
-                if (!ElementEquals(sourceEnumerator.Current, otherEnumerator.Current)) {
+                if (!Equals(sourceEnumerator.Current, otherEnumerator.Current)) {
                     // values aren't equal
                     return false;
                 }
@@ -88,7 +64,7 @@ namespace Exceptionless {
                     return false;
                 }
 
-                if (!ElementEquals(sourceValue, otherValue)) {
+                if (!Equals(sourceValue, otherValue)) {
                     return false;
                 }
             }
